Moved Markets Module to Skyze Std Lib
