+++ Skyze_Indicators_Library/Unit_Test/MovingAverage_test.py
@@ -6,7 +6,7 @@
 # 3rd Party
 
 # Our libraries - Required
-from Market import Market
+from Skyze_Standard_Library.Market import Market
 
 # Skyze imports
 from Skyze_Standard_Library.Unit_Test.UnitTestSkyzeAbstract import *       # Parent import
