+++ StrategySTC.py
@@ -10,7 +10,7 @@ import datetime
 # Our System
 from StrategyAbstract import StrategyAbstract
 from Portfolio import Portfolio
-from Market import Market
+from Skyze_Standard_Library.Market import Market
 from MarketStrategy import MarketStrategy
 from BackTester import BackTester
